+++ templates/parts/admin/welcome-promo.php
@@ -7,7 +7,8 @@
  * @package    Michelle
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.0.11
  */
 
 namespace WebManDesign\Michelle;
@@ -25,13 +26,13 @@ if ( ! class_exists( 'WebManDesign\Michelle\Welcome\Component' ) ) {
 
 <div class="about__section is-feature welcome--promo">
 	<h2>
-		<?php esc_html_e( 'Fancy something more?', 'michelle' ); ?>
+		<?php esc_html_e( 'Like the theme?', 'michelle' ); ?>
 	</h2>
 	<p>
-		<?php esc_html_e( 'Get priority support, more customization options and editor palette colors, sidebars and widgets, true one click demo content import, and even more by purchasing one of my paid themes.', 'michelle' ); ?>
-		<br>
-		<small><em><?php esc_html_e( '(Note that the theme you are using now is fully functional and free without any paid upgrade.)', 'michelle' ); ?></em></small>
+		<?php esc_html_e( 'You are using a fully functional 100% free WordPress theme without any paid upgrade.', 'michelle' ); ?>
+		<?php esc_html_e( 'If you find it helpful, please support its updates and technical support service with a donation or by purchasing one of paid products at WebManDesign.eu.', 'michelle' ); ?>
+		<?php esc_html_e( 'Thank you!', 'michelle' ); ?>
 		<br><br>
-		<a href="https://www.webmandesign.eu/?utm_source=michelle" class="button button-hero button-primary welcome--button-outline"><?php esc_html_e( 'Visit WebMan Design website now &rarr;', 'michelle' ); ?></a>
+		<a href="https://www.webmandesign.eu/contact/#donation" class="button button-hero button-primary welcome--button-outline"><?php esc_html_e( 'Visit WebMan Design website now &rarr;', 'michelle' ); ?></a>
 	</p>
 </div>